backend/requirements.txt
fastapi
uvicorn
sentence-transformers
sqlalchemy
gpt-researcher
qdrant_client
langgraph
ollama
langchain_google_vertexai
python-dotenv
vertexai
google-genai